requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

asgiref==3.4.1 ; python_version >= '3.6'
backports.entry-points-selectable==1.1.1 ; python_version >= '2.7'
beautifulsoup4==4.10.0 ; python_version >= '3.1'
certifi==2021.10.8
charset-normalizer==2.0.9 ; python_version >= '3'
cssutils==2.3.0 ; python_version >= '3.6'
distlib==0.3.3
django-dotenv==1.4.2
django-inlinecss==0.3.0
django-seed==0.3.1
django-storages===1.12.3
django-imagekit==4.1.0
django==3.2.5
faker==9.9.0 ; python_version >= '3.6'
filelock==3.4.0 ; python_version >= '3.6'
flake8==4.0.1
future==0.18.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3 ; python_version >= '3'
mccabe==0.6.1
pbr==5.8.0 ; python_version >= '2.6'
pillow==8.4.0
pip==21.3.1 ; python_version >= '3.6'
pipenv-to-requirements==0.9.0
pipenv==2021.11.23 ; python_version >= '3.6'
platformdirs==2.4.0 ; python_version >= '3.6'
pycodestyle==2.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyflakes==2.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pynliner==0.8.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
requests==2.26.0
setuptools==59.5.0 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.1 ; python_version >= '3.6'
sqlparse==0.4.2 ; python_version >= '3.5'
text-unidecode==1.3
toposort==1.7
psycopg2-binary==2.9.2
sentry-sdk==1.5.0
pyyaml==5.4.1
gunicorn==20.1.0
whitenoise==5.3.0
django-cleanup==5.2.0
urllib3==1.26.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
virtualenv-clone==0.5.7 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.10.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
